+++ lib/twinturbo/controller.rb
@@ -29,7 +29,7 @@ module TwinTurbo
         rescue NameError
         end
       end
-      puts "Could not find permitter: #{permitter_class_arr.collect{|c|c.to_s}.join(', ')}" if RestfulJson.debug?
+      logger.debug "Could not find permitter: #{permitter_class_arr.collect{|c|c.to_s}.join(', ')}" if RestfulJson.debug?
       nil
     end