+++ spec/spec_helper.rb
@@ -22,6 +22,8 @@ require 'rspec/rails'
 require 'json_spec'
 require 'restful_json'
 
+RestfulJson.debug = true
+
 Rails.backtrace_cleaner.remove_silencers!
 
 Dir["#{File.dirname(__FILE__)}/support/**/*.rb"].each { |f| require f }